+++ src/main/java/software/xdev/pmd/config/PluginConfiguration.java
@@ -13,7 +13,7 @@ import org.jetbrains.annotations.Nullable;
 
 import software.xdev.pmd.config.plugin.PatternContainer;
 import software.xdev.pmd.config.plugin.ThirdPartyClasspathConfigContainer;
-import software.xdev.pmd.model.config.ConfigurationLocation;
+import software.xdev.pmd.model.config.rulesetlocation.ConfigurationLocation;
 import software.xdev.pmd.model.scope.ScanScope;
