+++ src/main/java/software/xdev/pmd/config/PluginConfigurationBuilder.java
@@ -14,9 +14,9 @@ import com.intellij.openapi.project.Project;
 
 import software.xdev.pmd.config.plugin.PatternContainer;
 import software.xdev.pmd.config.plugin.ThirdPartyClasspathConfigContainer;
-import software.xdev.pmd.model.config.ConfigurationLocation;
-import software.xdev.pmd.model.config.ConfigurationLocationFactory;
-import software.xdev.pmd.model.config.bundled.BundledConfig;
+import software.xdev.pmd.model.config.rulesetlocation.ConfigurationLocation;
+import software.xdev.pmd.model.config.rulesetlocation.ConfigurationLocationFactory;
+import software.xdev.pmd.model.config.rulesetlocation.bundled.BundledConfig;
 import software.xdev.pmd.model.scope.ScanScope;
