+++ src/main/resources/META-INF/plugin.xml
@@ -25,7 +25,7 @@
 	<extensions defaultExtensionNs="com.intellij">
 		<projectService serviceImplementation="software.xdev.pmd.analysis.PMDAnalyzer"/>
 		<projectService serviceImplementation="software.xdev.pmd.analysis.validate.PsiFileValidator"/>
-		<projectService serviceImplementation="software.xdev.pmd.model.config.ConfigurationLocationFactory"/>
+		<projectService serviceImplementation="software.xdev.pmd.model.config.rulesetlocation.ConfigurationLocationFactory"/>
 		<projectService serviceImplementation="software.xdev.pmd.config.ConfigurationLocationSource"/>
 		<projectService serviceImplementation="software.xdev.pmd.config.PluginConfigurationManager"/>
 		<projectService serviceImplementation="software.xdev.pmd.analysis.ProjectScanClasspathManager"/>
